requirements.txt
snscrape>=0.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0
rich>=13.0.0
click>=8.1.0
python-dotenv>=1.0.0
ollama>=0.1.0
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0
colorama>=0.4.6
tqdm>=4.65.0
schedule>=1.2.0
